cli/sql: split internal tests into a new file
"internal" tests are those that are interested in the internals of the
SQL shell. The remaining tests in `sql_test.go` only use the external
CLI.

Release note: None
